Removed interface AuthorityManager
