+++ webcurator-core/src/test/java/org/webcurator/core/admin/MockPermissionTemplateManagerImpl.java
@@ -2,14 +2,12 @@ package org.webcurator.core.admin;
 
 import org.webcurator.auth.*;
 import org.webcurator.domain.*;
-import org.webcurator.domain.model.auth.User;
-import org.webcurator.domain.model.core.PermissionTemplate;
 
 public class MockPermissionTemplateManagerImpl extends PermissionTemplateManagerImpl {
 
 	public MockPermissionTemplateManagerImpl(String filename) {
 		super();
-		this.setAuthorityManager(new AuthorityManagerImpl());
+		this.setAuthorityManager(new AuthorityManager());
 		this.setPermissionTemplateDAO(new MockPermissionTemplateDAO(filename));
 	}
