+++ webcurator-core/src/test/java/org/webcurator/core/targets/MembersRemovedEventPropagatorTest.java
@@ -4,18 +4,7 @@ import static org.junit.Assert.*;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.springframework.context.MockMessageSource;
-import org.webcurator.auth.AuthorityManagerImpl;
-import org.webcurator.core.notification.MockInTrayManager;
-import org.webcurator.core.scheduler.MockTargetInstanceManager;
 import org.webcurator.core.scheduler.TargetInstanceManager;
-import org.webcurator.core.util.TestAuditor;
-import org.webcurator.domain.MockAnnotationDAO;
-import org.webcurator.domain.MockSiteDAO;
-import org.webcurator.domain.MockTargetDAO;
-import org.webcurator.domain.MockTargetInstanceDAO;
-import org.webcurator.domain.model.core.BusinessObjectFactory;
-import org.webcurator.domain.model.core.Target;
 import org.webcurator.domain.model.core.TargetGroup;
 import org.webcurator.domain.model.core.TargetInstance;
 import org.webcurator.test.BaseWCTTest;
