+++ webcurator-core/src/test/java/org/webcurator/core/targets/MockTargetManager.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.context.MockMessageSource;
-import org.webcurator.auth.AuthorityManagerImpl;
+import org.webcurator.auth.AuthorityManager;
 import org.webcurator.core.common.WCTTreeSet;
 import org.webcurator.core.notification.MockInTrayManager;
 import org.webcurator.core.scheduler.MockTargetInstanceManager;
@@ -22,7 +22,7 @@ public class MockTargetManager extends TargetManagerImpl {
 		this.setSiteDao(new MockSiteDAO(fileName));
 		this.setAnnotationDAO(new MockAnnotationDAO(fileName));
 		this.setTargetInstanceDao(new MockTargetInstanceDAO(fileName));
-		this.setAuthMgr(new AuthorityManagerImpl());
+		this.setAuthMgr(new AuthorityManager());
 		this.setInstanceManager(new MockTargetInstanceManager(fileName));
 		this.setIntrayManager(new MockInTrayManager(fileName));
 		this.setAuditor(new TestAuditor());
